# [CHIA-1087] validate blocks in thread pool (instead of process pool)