--- a/chia/_tests/core/mempool/test_mempool_manager.py
+++ b/chia/_tests/core/mempool/test_mempool_manager.py
@@ -227,6 +227,7 @@ def make_test_conds(
         cost,
         0,
         0,
+        False,
     )