This PR is fairly big. It has been split up into smaller, incremental, commits to make it easier to review.

### Purpose:

The overarching goal is to simplify the code and improve block validation performance/efficiency.

Moving block validation into a thread pool (instead of a process pool, as it is today) has the following benefits:

* worker processes can sometimes be left behind when not shutting down gracefully. Threads don't have this problem
* In order to execute the work in a separate process, we need to copy any inputs over IPC. With threads, we can use shared memory and avoid the copies.
* As more of the block validation moves into rust, we can keep the GIL unlocked for longer periods at a time, improving concurrency

The main risk of moving to threads is that the GIL will be required so often that it will be a net loss in performance. As long as `run_block_generator()`, `AugScheme.aggregate_verify()` and `Verifier.validate_proof()` all release the GIL, the bulk of the CPU time is spent with it released. In this case the performance seems mostly unaffected by this change.

~~We need a new `chia_rs` release for the signature validation to release the GIL.~~ `chia_rs` has been updated.

### Current Behavior:

Block validation happens in a `ProcessPoolExecutor`, and blocks, previous generators and previous block records are copied over to the worker process.

### New Behavior:

Block validation happens in a `ThreadPoolExecutor`, and blocks, previous generators and previous block records are shared between the threads (avoiding a copy).

### Testing Notes:

The coverage diff of the *exception* failure case did not have any tests, but is possibly also defensive programming. I think it's fine to keep those paths uncovered. Perhaps at some point in the future we'll unify our error handling to use exceptions everywhere.

#### Performance

To ensure this did not materially impact performance, I ran the following test:

```
pytest -n 0 -s chia/_tests/core/full_node/test_full_node.py::test_long_reorg
```
| test | before | after | after/before |
| --- | --- | --- |--- |
| `test_full_node.py::test_long_reorg` | 1024.80s| 1032.20s | 1.0072 |

running the full sync benchmarks, this patch causes a slow-down until we have an updated `chia_rs` release.

| test | main | this PR | this PR with latest chia_rs |
| --- | --- | --- | --- |
| sync full | 200.93 s | 289.40 s | 173.54 s |
| sync empty | 15.45 s | 33.53 s | 15.63 s |
| keepup full | 587.32 s | 589.88 s | 595.98 s |

#### Sync test

There's a sync test (requires VPN) [here](https://grafana.sin.chiatechlab.com/d/oBgwkDI7z/blockchain-sync-tests?orgId=1&var-ref=837f0e2d472cdec45d4eb880b908b43af8b58d5d&var-ref=validate-thread-pool&var-repo=All&from=now-3d&to=now&var-network=mainnet), suggesting this patch gives slightly higher performance/efficiency.